refactor: move useHeightResizeObserver to ui-kit, add docs and options parameter
