+++ apps/main/src/dao/layout/index.tsx
@@ -1,13 +1,13 @@
+import Header from '@/dao/layout/Header'
+import useStore from '@/dao/store/useStore'
+import { getNetworkFromUrl } from '@/dao/utils/utilsRouter'
+import { getWalletChainId, useWallet } from '@ui-kit/features/connect-wallet'
+import useHeightResizeObserver from '@ui-kit/hooks/useHeightResizeObserver'
+import { isChinese, t } from '@ui-kit/lib/i18n'
+import { Footer } from '@ui-kit/widgets/Footer'
+import { CONNECT_STAGE, isFailure, isLoading } from '@ui/utils'
 import { ReactNode, useEffect, useMemo, useRef, useState } from 'react'
 import styled from 'styled-components'
-import { isChinese, t } from '@ui-kit/lib/i18n'
-import { CONNECT_STAGE, isFailure, isLoading } from '@ui/utils'
-import { getWalletChainId, useWallet } from '@ui-kit/features/connect-wallet'
-import { getNetworkFromUrl } from '@/dao/utils/utilsRouter'
-import { useHeightResizeObserver } from '@ui/hooks'
-import useStore from '@/dao/store/useStore'
-import Header from '@/dao/layout/Header'
-import { Footer } from '@ui-kit/widgets/Footer'
 
 const BaseLayout = ({ children }: { children: ReactNode }) => {
   const { wallet } = useWallet()
