+++ apps/main/src/lend/layout/index.tsx
@@ -1,15 +1,15 @@
+import { CONNECT_STAGE, ROUTE } from '@/lend/constants'
+import Header from '@/lend/layout/Header'
+import { layoutHeightKeys } from '@/lend/store/createLayoutSlice'
+import useStore from '@/lend/store/useStore'
+import { getNetworkFromUrl } from '@/lend/utils/utilsRouter'
+import { getWalletChainId, useWallet } from '@ui-kit/features/connect-wallet'
+import useHeightResizeObserver from '@ui-kit/hooks/useHeightResizeObserver'
+import { isChinese, t } from '@ui-kit/lib/i18n'
+import { Footer } from '@ui-kit/widgets/Footer'
+import { isFailure, isLoading } from '@ui/utils'
 import { ReactNode, useEffect, useMemo, useRef, useState } from 'react'
 import styled from 'styled-components'
-import { CONNECT_STAGE, ROUTE } from '@/lend/constants'
-import { layoutHeightKeys } from '@/lend/store/createLayoutSlice'
-import { getNetworkFromUrl } from '@/lend/utils/utilsRouter'
-import { isFailure, isLoading } from '@ui/utils'
-import { getWalletChainId, useWallet } from '@ui-kit/features/connect-wallet'
-import useStore from '@/lend/store/useStore'
-import Header from '@/lend/layout/Header'
-import { Footer } from '@ui-kit/widgets/Footer'
-import { useHeightResizeObserver } from '@ui/hooks'
-import { isChinese, t } from '@ui-kit/lib/i18n'
 
 const BaseLayout = ({ children }: { children: ReactNode }) => {
   const { wallet } = useWallet()
