+++ packages/ui/src/hooks/index.ts
@@ -1,4 +1,3 @@
-export { default as useHeightResizeObserver } from './useHeightResizeObserver'
 export { default as useIntersectionObserver } from './useIntersectionObserver'
 export { default as usePageVisibleInterval } from './usePageVisibleInterval'
 export { default as useSlideTabState } from './useSlideTabState'
